// frontend/src/types/analytics.ts
export type LocationData = {
  count: number;
  name: string;
  coords: [number, number];
  key: string;
};